refactor: Renamed classes and folder
