+++ serverless/database-backup/src/dumps/secrets-manager-dump.ts
