+++ serverless/database-backup/src/index.ts
@@ -2,9 +2,9 @@ import 'source-map-support/register'
 import * as Sentry from '@sentry/node'
 
 import config from './config'
-import { PgDump, SecretsManagerDump } from './backups'
+import { PgDump, SecretsManagerDump } from './dumps'
 import Encryptor from './encryptor'
-import Upload from './upload'
+import Backup from './backup'
 
 import { EncryptionConfig, DatabaseConfig } from './interfaces'
 import { parseRdsEvents, RDS_EVENTS } from './utils/event'
@@ -30,13 +30,17 @@ const handler = async (event: any) => {
     for (const ev of events) {
       if (ev.eventId === RDS_EVENTS.BACKUP_COMPLETE) {
         const dbConfig = config.get('database') as DatabaseConfig
-        const pgdump = new PgDump(dbConfig)
-        const secretsManagerDump = new SecretsManagerDump(dbConfig)
-
         const encryptionConfig = config.get('encryption') as EncryptionConfig
+
+        const pgDump = new PgDump(dbConfig)
+        const secretsDump = new SecretsManagerDump(dbConfig)
         const encryptor = new Encryptor(encryptionConfig)
 
-        const backup = new Upload(encryptor, pgdump, secretsManagerDump)
+        const backup = new Backup({
+          encryptor,
+          pgDump,
+          secretsDump,
+        })
         const backupLocation = await backup.upload()
 
         console.log(`Database backup uploaded to ${backupLocation}`)
